+++ hal/port/unix/lt_port_unix_spi.c
@@ -160,9 +160,9 @@ lt_ret_t lt_port_spi_csn_high(lt_l2_state_t *s2)
     return LT_OK;
 }
 
-lt_ret_t lt_port_spi_transfer(lt_l2_state_t *s2, uint8_t offset, uint16_t tx_data_length, uint32_t timeout)
+lt_ret_t lt_port_spi_transfer(lt_l2_state_t *s2, uint8_t offset, uint16_t tx_data_length, uint32_t timeout_ms)
 {
-    UNUSED(timeout);
+    UNUSED(timeout_ms);
     lt_dev_unix_spi_t *device = (lt_dev_unix_spi_t *)(s2->device);
 
     int ret = 0;