+++ hal/port/unix/lt_port_unix_tcp.c
@@ -240,9 +240,9 @@ lt_ret_t lt_port_spi_csn_high(lt_l2_state_t *s2)
     return communicate(dev, NULL, NULL);
 }
 
-lt_ret_t lt_port_spi_transfer(lt_l2_state_t *s2, uint8_t offset, uint16_t tx_data_length, uint32_t timeout)
+lt_ret_t lt_port_spi_transfer(lt_l2_state_t *s2, uint8_t offset, uint16_t tx_data_length, uint32_t timeout_ms)
 {
-    UNUSED(timeout);
+    UNUSED(timeout_ms);
     lt_dev_unix_tcp_t *dev = (lt_dev_unix_tcp_t *)(s2->device);
     lt_ret_t ret;