+++ src/lt_l1.h
@@ -46,10 +46,10 @@
  *
  * @param s2          Structure holding l2 state
  * @param max_len     Max len of receive buffer
- * @param timeout     Timeout - how long function will wait for response
+ * @param timeout_ms  Timeout - how long function will wait for response
  * @return            LT_OK if success, otherwise returns other error code.
  */
-lt_ret_t lt_l1_read(lt_l2_state_t *s2, const uint32_t max_len, const uint32_t timeout)
+lt_ret_t lt_l1_read(lt_l2_state_t *s2, const uint32_t max_len, const uint32_t timeout_ms)
     __attribute__((warn_unused_result));
 
 /**
@@ -57,10 +57,11 @@ lt_ret_t lt_l1_read(lt_l2_state_t *s2, const uint32_t max_len, const uint32_t ti
  *
  * @param s2          Structure holding l2 state
  * @param len         Length of data to send
- * @param timeout     Timeout
+ * @param timeout_ms  Timeout
  * @return            LT_OK if success, otherwise returns other error code.
  */
-lt_ret_t lt_l1_write(lt_l2_state_t *s2, const uint16_t len, const uint32_t timeout) __attribute__((warn_unused_result));
+lt_ret_t lt_l1_write(lt_l2_state_t *s2, const uint16_t len, const uint32_t timeout_ms)
+    __attribute__((warn_unused_result));
 
 /** @} */  // end of group_l1_functions